[dbsp] Rearrange `storage::file` module a little.
This moves `storage::file::reader` into a subdirectory, to prepare for
adding some more submodules, and moves `storage::file::test` into its own
file.
